Separate storage test in e2e tests
- Add suite into the `test-storage` make target
- Add new package to host e2e helper funcs
- Add new package to host e2e storage test
- Move storage tests into new package `storage_kube_test`
