+++ e2e/kube/examples_test.go
@@ -156,26 +156,6 @@ var _ = Describe("Examples", func() {
 				Expect(string(outSecretDecoded)).To(Equal(strings.TrimSuffix(outFile, "\n")))
 			})
 
-			It("bosh-deployment with a persistent disk example must work", func() {
-				yamlFilePath := examplesDir + "bosh-deployment/boshdeployment-with-persistent-disk.yaml"
-
-				By("Creating bosh deployment")
-				kubectlHelper := testing.NewKubectl()
-				err := kubectlHelper.Create(namespace, yamlFilePath)
-				Expect(err).ToNot(HaveOccurred())
-
-				By("Checking for pods")
-				err = kubectlHelper.Wait(namespace, "ready", "pod/nats-deployment-nats-v1-0")
-				Expect(err).ToNot(HaveOccurred())
-
-				err = kubectlHelper.Wait(namespace, "ready", "pod/nats-deployment-nats-v1-1")
-				Expect(err).ToNot(HaveOccurred())
-
-				By("Checking for pvc")
-				err = kubectlHelper.WaitForPVC(namespace, "nats-deployment-nats-pvc")
-				Expect(err).ToNot(HaveOccurred())
-			})
-
 			It("extended-job auto errand delete example must work", func() {
 				yamlFilePath := examplesDir + "extended-job/exjob_auto-errand-deletes-pod.yaml"
